fix build break on linux/mac which are more restrictive in where they define size_t in the headers

+++ shared/ebm_native/inc/ebm_native.h
@@ -84,11 +84,11 @@ extern "C" {
 #endif // compiler type
 
 typedef struct _BoosterHandle {
-   size_t handleVerification; // should be 25077 if ok
+   uint32_t handleVerification; // should be 25077 if ok
 } * BoosterHandle;
 
 typedef struct _InteractionHandle {
-   size_t handleVerification; // should be 27917 if ok
+   uint32_t handleVerification; // should be 27917 if ok
 } * InteractionHandle;
 
 #ifndef PRId32
